#149 Aggregation operators moved
