+++ libakumuli/CMakeLists.txt
@@ -16,6 +16,7 @@ add_library(akumuli SHARED
     storage_engine/compression.cpp
     storage_engine/column_store.cpp
     storage_engine/operators/operator.cpp
+    storage_engine/operators/aggregate.cpp
     storage_engine/operators/scan.cpp
     # query_processing
     query_processing/queryparser.cpp
