+++ unittests/CMakeLists.txt
@@ -185,6 +185,7 @@ add_executable(
     ../libakumuli/storage_engine/nbtree.cpp
     ../libakumuli/storage_engine/compression.cpp
     ../libakumuli/storage_engine/operators/operator.cpp
+    ../libakumuli/storage_engine/operators/aggregate.cpp
     ../libakumuli/storage_engine/operators/scan.cpp
     ../libakumuli/storage_engine/column_store.cpp
     ../libakumuli/query_processing/queryparser.cpp
@@ -339,6 +340,7 @@ add_executable(
     ../libakumuli/storage_engine/nbtree.cpp
     ../libakumuli/storage_engine/compression.cpp
     ../libakumuli/storage_engine/operators/operator.cpp
+    ../libakumuli/storage_engine/operators/aggregate.cpp
     ../libakumuli/storage_engine/operators/scan.cpp
     ../libakumuli/util.cpp
     ../libakumuli/status_util.cpp
@@ -366,6 +368,7 @@ add_executable(
     ../libakumuli/storage_engine/nbtree.cpp
     ../libakumuli/storage_engine/compression.cpp
     ../libakumuli/storage_engine/operators/operator.cpp
+    ../libakumuli/storage_engine/operators/aggregate.cpp
     ../libakumuli/storage_engine/operators/scan.cpp
     ../libakumuli/storage_engine/column_store.cpp
     ../libakumuli/util.cpp
